Pytest: Improve doctest integration
