+++ .gitignore
@@ -18,7 +18,6 @@
 /config.log
 /config.status
 /configure
-/conftest*
 /confdefs.h
 
 /m4/sage_spkg_configures.m4
