+++ src/tox.ini
@@ -376,14 +376,6 @@ exclude =
     sage/graphs/graph_plot.py
     sage/misc/sagedoc.py
 
-[pytest]
-python_files = *_test.py
-norecursedirs = local prefix venv build builddir pkgs .git src/doc src/bin tools
-addopts = --import-mode importlib
-doctest_optionflags = NORMALIZE_WHITESPACE ELLIPSIS
-# https://docs.pytest.org/en/stable/reference/reference.html#confval-consider_namespace_packages
-consider_namespace_packages = True
-
 [coverage:run]
 source = sage
 concurrency = multiprocessing,thread
